Commit sasamitime hotfixes made in production to git
I had to rush to upload sasamitime before I left anime society at 22:00.
As a result, there was left-over code and broken time handling. After leaving,
I was on the move and so couldn't get out my laptop and update the code
properly, so I had to SSH into the server on my phone and edit with vim, live,
saving and refreshing the production site to see the results. Now that I'm on
the bus home, I am able to get out my laptop again and finally commit the
changes I made.

Things changed:
- Fixed time handling to apply modulo and division in the correct order
- Fixed time handling to acknowledge that clockfaces show TWELVE-hour time, not
  twenty-four-hour time
- Removed commented-out code
- Removed now-unnecessary code
- Fixed angle hands are drawn at